+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- [Kiq.Client.Resurrection] Correct usage of SCAN to ensure all backup queues
+  will be found for re-enqueueing.
+
 ## [0.5.0] — 2018-12-14
 
 ### Added