+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.7.3] — 2023-07-18
+
+- Update dependency constraints for `telemetry`, `nimble_parsec`, `redix`, and `gen_stage`
+
 ## [0.7.2] — 2021-05-12
 
 - Update and loosen dependency constraints